Refactor ObjectStoragePath to airflow.io.path
