+++ docs/apache-airflow/core-concepts/objectstorage.rst
@@ -70,7 +70,7 @@ object you want to interact with. For example, to point to a bucket in s3, you w
 
 .. code-block:: python
 
-    from airflow.io.store.path import ObjectStoragePath
+    from airflow.io.path import ObjectStoragePath
 
     base = ObjectStoragePath("s3://my-bucket/", conn_id="aws_default")  # conn_id is optional
 
@@ -149,7 +149,7 @@ would do the following:
 
 .. code-block:: python
 
-    from airflow.io.store.path import ObjectStoragePath
+    from airflow.io.path import ObjectStoragePath
     from airflow.io.store import attach
 
     from fsspec.implementations.dbfs import DBFSFileSystem
@@ -172,7 +172,7 @@ The object storage abstraction is implemented as a `Path-like API <https://docs.
 This means that you can mostly use the same API to interact with object storage as you would with a local filesystem.
 In this section we only list the differences between the two APIs. Extended operations beyond the standard Path API
 , like copying and moving, are listed in the next section. For details about each operation, like what arguments
-they take, see the documentation of the :class:`~airflow.io.store.path.ObjectStoragePath` class.
+they take, see the documentation of the :class:`~airflow.io.path.ObjectStoragePath` class.
 
 
 stat
@@ -291,7 +291,7 @@ are used to connect to s3 and a parquet file, indicated by a ``ObjectStoragePath
 .. code-block:: python
 
     import duckdb
-    from airflow.io.store.path import ObjectStoragePath
+    from airflow.io.path import ObjectStoragePath
 
     path = ObjectStoragePath("s3://my-bucket/my-table.parquet", conn_id="aws_default")
     conn = duckdb.connect(database=":memory:")
