+++ tests/system/providers/common/io/example_file_transfer_local_to_s3.py
@@ -23,7 +23,7 @@ from typing import cast
 
 from airflow import DAG
 from airflow.decorators import task
-from airflow.io.store.path import ObjectStoragePath
+from airflow.io.path import ObjectStoragePath
 from airflow.providers.common.io.operators.file_transfer import FileTransferOperator
 from airflow.utils.trigger_rule import TriggerRule
